completion DashboardService

+++ src/main/java/com/example/wallet_management_app/entity/AlertSetting.java
@@ -11,6 +11,8 @@ import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import jakarta.persistence.FetchType;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -45,5 +47,5 @@ public class AlertSetting extends BaseEntity {
     private Boolean remainingAmountAlertEnabled;
 
     @Column(name = "remaining_amount_threshold", nullable = false)
-    private Long remainingAmountThreshold;
+    private BigDecimal remainingAmountThreshold;
 }
